# docs/openapi-v1.yml
openapi: 3.0.3
info:
  title: Selene API — v1
  description: >
    Public REST contract for the Selene smart energy & climate platform.
    Served by the Caddy gateway under the `/api/v1` prefix; each path below is
    routed to a dedicated microservice (services/auth, services/energy,
    services/climate, services/analytics, services/firmware).

    Versioning lives at the gateway: `handle_path /api/v1/*` strips the prefix
    before proxying, so services define version-agnostic paths (`/energy/latest`)
    and a future `/api/v2` needs no service changes.

    Auth: short-lived EdDSA (Ed25519) access tokens with rotating refresh
    tokens (one-time use, reuse detection). JWKS at `/.well-known/jwks.json`.
  version: 1.0.0
  contact:
    name: Selene
  license:
    name: Proprietary
    url: https://github.com/dankehidayat/selene
servers:
  - url: https://selene.dankehidayat.my.id/api/v1
    description: Production
  - url: http://localhost:3009/api/v1
    description: Local (gateway / auth)
...
paths:
...
  /admin/users:
    get:

      operationId: adminListUsers
      tags: [admin]
      summary: List users (admin only)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: role
          in: query
          schema: { type: string, enum: [USER, ADMIN] }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: Paginated list of users.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/UserProfile"
                  total:
                    type: integer
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/users/{id}/role:
    patch:

      operationId: adminChangeUserRole
      tags: [admin]
      summary: Change a user's role (admin only; requires 2FA when elevation)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          description: User id.
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [role]
              properties:
                role:
                  type: string
                  enum: [USER, ADMIN]
                totpCode:
                  type: string
                  description: Required when elevating to ADMIN (security spec).
      responses:
        "200":
          description: Updated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /admin/stats:
    get:
      operationId: adminStats
      tags: [admin]
      summary: Admin statistics dashboard
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Aggregated admin stats.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "403": { $ref: "#/components/responses/Forbidden" }
  /admin/users/{id}/toggle-active:
    patch:
      operationId: adminToggleUserActive
      tags: [admin]
      summary: Toggle a user's active status (admin only)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          description: User id.
          schema: { type: string }
      responses:
        "200":
          description: Updated user with toggled active status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "403": { $ref: "#/components/responses/Forbidden" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived EdDSA (Ed25519) access token; refresh via POST /auth/refresh. Verify against /.well-known/jwks.json.
...
  responses:
    BadRequest:
      description: Invalid payload.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Missing, invalid, or expired token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Valid token but insufficient permissions.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    RateLimited:
      description: Too many requests; honor the Retry-After header.
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              enum: [UNAUTHORIZED, FORBIDDEN, RATE_LIMITED, VALIDATION_ERROR, NOT_FOUND, CONFLICT, INTERNAL_ERROR, TOKEN_REUSE_DETECTED]
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            timestamp:
              type: string
              format: date-time
            request_id:
              type: string
...
    UserProfile:
      type: object
      additionalProperties: false
      required: [id, email, role]
      properties:
        id: { type: string }
        email: { type: string }
        name: { type: string, nullable: true }
        role: { type: string, enum: [USER, ADMIN] }
        totpEnabled: { type: boolean }
        createdAt: { type: string, format: date-time }
...
    AdminStats:
      type: object
      additionalProperties: false
      required: [totalUsers, activeUsers, adminUsers, totalLogins]
      properties:
        totalUsers:
          type: integer
        activeUsers:
          type: integer
        adminUsers:
          type: integer
        totalLogins:
          type: integer
...